refactor skills
