+++ src/pages/calendar/index.tsx
@@ -1,5 +1,5 @@
-import Head from "next/head";
 import React from "react";
+import Head from "next/head";
 import useStorage from "react-use-localstorage2";
 import { Calendar, Day, LoadQuestions, Month, Navbar } from "../../calendar";
