+++ src/pages/persona/index.tsx
@@ -1,5 +1,5 @@
-import Head from "next/head";
 import React from "react";
+import Head from "next/head";
 import { Personas } from "../../data";
 import { personaHeaders, PersonaRowParser } from "../../personaList";
 import PrepareTable from "../../table";
