--- src/pages/skill/[skill].ts
@@ -1,2 +1,0 @@
-import SkillInfo from "../../skill";
-export default SkillInfo;
