--- src/pages/skill/index.ts
@@ -1,2 +1,0 @@
-import SkillTable from "../../skillList";
-export default SkillTable;
