+++ src/skillList/SkillList.tsx
@@ -1,15 +1,14 @@
 import React from "react";
 import Link from "next/link";
-import { useRouter } from "next/router";
 import Image from "next/image";
-import { Row, Cell } from "react-table";
-import PrepareTable from "../table";
-import { Skills, Skill, PersonaRef } from "../data";
+import { Skill, PersonaRef } from "../data";
 import { SkillTags, CostTag, StatusTag } from "../skill";
 import { Elements } from "../persona";
+import router from "next/router";
+import { Row, Cell } from "react-table";
 
 const camel = (text: string) => text.charAt(0).toUpperCase() + text.slice(1);
-const Headers = () =>
+export const Headers = () =>
   React.useMemo(
     () => [
       {
@@ -76,10 +75,8 @@ const Headers = () =>
     []
   );
 
-export const SkillList = (): JSX.Element => {
-  const router = useRouter();
-  const headers = Headers();
-  const rowParser = ({ getRowProps, cells, values: { name } }: Row, index: number) => {
+
+export const SkillRowParser = ({ getRowProps, cells, values: { name } }: Row, index: number) => {
     const parseClick = ({ target }) => {
       if (target.tagName.toLowerCase() == "a") return false;
       router.push(`/skill/${name}`);
@@ -96,6 +93,3 @@ export const SkillList = (): JSX.Element => {
       </div>
     );
   };
-  return <PrepareTable {...{ columns: headers, data: Skills, rowParser, className: "skills", sortId: "name" }} />;
-};
-export default SkillList;
